separated the policy structs and methods into policy.go Added single policy get for rotation and dual auth policy update the unit tests for single policy get
